gemini_License_V3_gemini_structure/DineDashPOS/requirements.txt
customtkinter
Pillow
python-escpos
openpyxl
pandas
matplotlib